+++ changelog.md
@@ -154,6 +154,8 @@ More data saving and map activity features.
 * Document `App.js`
 * Add question state to `RandomAudioActivity`
 * Render form questions in random audio
+* Move activities to their own directory
+
 
 I used a couple blog posts about how to set up Express and postgres:
 * https://medium.freecodecamp.org/how-to-make-create-react-app-work-with-a-node-backend-api-7c5c48acb1b0